Remove old reflection destructurer, only new one based on fastmember remains
