+++ Tests/Serilog.Exceptions.Benchmark/BenchmarkDotNet.Artifacts/results/Serilog.Exceptions.Benchmark.DestructuringBenchmark-report-github.md
@@ -10,11 +10,9 @@ Frequency=2437499 Hz, Resolution=410.2566 ns, Timer=TSC
 
 
 ```
-|                     Method |  Job | Runtime |       Mean |      Error |      StdDev |     Median |
-|--------------------------- |----- |-------- |-----------:|-----------:|------------:|-----------:|
-|  OldReflectionDestructurer |  Clr |     Clr |  15.360 us |  0.2928 us |   0.3133 us |  15.412 us |
-|         CustomDestructurer |  Clr |     Clr |   6.643 us |  0.1304 us |   0.1395 us |   6.615 us |
-| FastReflectionDestructurer |  Clr |     Clr |  11.316 us |  0.2241 us |   0.2201 us |  11.364 us |
-|  OldReflectionDestructurer | Core |    Core | 807.670 us | 20.8714 us |  51.1980 us | 792.939 us |
-|         CustomDestructurer | Core |    Core | 785.605 us | 10.6901 us |   9.9995 us | 788.961 us |
-| FastReflectionDestructurer | Core |    Core | 883.007 us | 42.0822 us | 124.0805 us | 813.333 us |
+|                 Method |  Job | Runtime |       Mean |      Error |     StdDev |
+|----------------------- |----- |-------- |-----------:|-----------:|-----------:|
+| ReflectionDestructurer |  Clr |     Clr |  11.180 us |  0.1812 us |  0.1415 us |
+|     CustomDestructurer |  Clr |     Clr |   6.408 us |  0.0514 us |  0.0456 us |
+| ReflectionDestructurer | Core |    Core | 742.092 us |  7.2303 us |  6.0376 us |
+|     CustomDestructurer | Core |    Core | 750.728 us | 18.9935 us | 34.2492 us |
