+++ Tests/Serilog.Exceptions.Benchmark/BenchmarkDotNet.Artifacts/results/Serilog.Exceptions.Benchmark.DestructuringBenchmark-report.csv
@@ -1,7 +1,5 @@
-Method;Job;AnalyzeLaunchVariance;EvaluateOverhead;MaxAbsoluteError;MaxRelativeError;MinInvokeCount;MinIterationTime;RemoveOutliers;Affinity;Jit;Platform;Runtime;AllowVeryLargeObjects;Concurrent;CpuGroups;Force;HeapAffinitizeMask;HeapCount;NoAffinitize;RetainVm;Server;Arguments;BuildConfiguration;Clock;EngineFactory;EnvironmentVariables;Toolchain;IsBaseline;InvocationCount;IterationTime;LaunchCount;RunStrategy;TargetCount;UnrollFactor;WarmupCount;Mean;Error;StdDev;Median
-OldReflectionDestructurer;Clr;False;Default;Default;Default;Default;Default;Default;15;RyuJit;X64;Clr;False;True;False;True;Default;Default;False;False;False;Default;Default;Default;Default;Default;Default;Default;1;Default;Default;Default;Default;16;Default;15.360 us;0.2928 us;0.3133 us;15.412 us
-CustomDestructurer;Clr;False;Default;Default;Default;Default;Default;Default;15;RyuJit;X64;Clr;False;True;False;True;Default;Default;False;False;False;Default;Default;Default;Default;Default;Default;Default;1;Default;Default;Default;Default;16;Default;6.643 us;0.1304 us;0.1395 us;6.615 us
-FastReflectionDestructurer;Clr;False;Default;Default;Default;Default;Default;Default;15;RyuJit;X64;Clr;False;True;False;True;Default;Default;False;False;False;Default;Default;Default;Default;Default;Default;Default;1;Default;Default;Default;Default;16;Default;11.316 us;0.2241 us;0.2201 us;11.364 us
-OldReflectionDestructurer;Core;False;Default;Default;Default;Default;Default;Default;15;RyuJit;X64;Core;False;True;False;True;Default;Default;False;False;False;Default;Default;Default;Default;Default;Default;Default;1;Default;Default;Default;Default;16;Default;807.670 us;20.8714 us;51.1980 us;792.939 us
-CustomDestructurer;Core;False;Default;Default;Default;Default;Default;Default;15;RyuJit;X64;Core;False;True;False;True;Default;Default;False;False;False;Default;Default;Default;Default;Default;Default;Default;1;Default;Default;Default;Default;16;Default;785.605 us;10.6901 us;9.9995 us;788.961 us
-FastReflectionDestructurer;Core;False;Default;Default;Default;Default;Default;Default;15;RyuJit;X64;Core;False;True;False;True;Default;Default;False;False;False;Default;Default;Default;Default;Default;Default;Default;1;Default;Default;Default;Default;16;Default;883.007 us;42.0822 us;124.0805 us;813.333 us
+Method;Job;AnalyzeLaunchVariance;EvaluateOverhead;MaxAbsoluteError;MaxRelativeError;MinInvokeCount;MinIterationTime;RemoveOutliers;Affinity;Jit;Platform;Runtime;AllowVeryLargeObjects;Concurrent;CpuGroups;Force;HeapAffinitizeMask;HeapCount;NoAffinitize;RetainVm;Server;Arguments;BuildConfiguration;Clock;EngineFactory;EnvironmentVariables;Toolchain;IsBaseline;InvocationCount;IterationTime;LaunchCount;RunStrategy;TargetCount;UnrollFactor;WarmupCount;Mean;Error;StdDev
+ReflectionDestructurer;Clr;False;Default;Default;Default;Default;Default;Default;15;RyuJit;X64;Clr;False;True;False;True;Default;Default;False;False;False;Default;Default;Default;Default;Default;Default;Default;1;Default;Default;Default;Default;16;Default;11.180 us;0.1812 us;0.1415 us
+CustomDestructurer;Clr;False;Default;Default;Default;Default;Default;Default;15;RyuJit;X64;Clr;False;True;False;True;Default;Default;False;False;False;Default;Default;Default;Default;Default;Default;Default;1;Default;Default;Default;Default;16;Default;6.408 us;0.0514 us;0.0456 us
+ReflectionDestructurer;Core;False;Default;Default;Default;Default;Default;Default;15;RyuJit;X64;Core;False;True;False;True;Default;Default;False;False;False;Default;Default;Default;Default;Default;Default;Default;1;Default;Default;Default;Default;16;Default;742.092 us;7.2303 us;6.0376 us
+CustomDestructurer;Core;False;Default;Default;Default;Default;Default;Default;15;RyuJit;X64;Core;False;True;False;True;Default;Default;False;False;False;Default;Default;Default;Default;Default;Default;Default;1;Default;Default;Default;Default;16;Default;750.728 us;18.9935 us;34.2492 us
