+++ Tests/Serilog.Exceptions.Benchmark/BenchmarkDotNet.Artifacts/results/Serilog.Exceptions.Benchmark.DestructuringBenchmark-report.html
@@ -24,14 +24,12 @@ Frequency=2437499 Hz, Resolution=410.2566 ns, Timer=TSC
 <pre><code></code></pre>
 
 <table>
-<thead><tr><th>              Method</th><th>Job</th><th>Runtime</th><th>Mean</th><th>Error</th><th>StdDev</th><th>Median</th>
+<thead><tr><th>          Method</th><th>Job</th><th>Runtime</th><th>Mean</th><th>Error</th><th>StdDev</th>
 </tr>
-</thead><tbody><tr><td>OldReflectionDestructurer</td><td>Clr</td><td>Clr</td><td>15.360 us</td><td>0.2928 us</td><td>0.3133 us</td><td>15.412 us</td>
-</tr><tr><td>CustomDestructurer</td><td>Clr</td><td>Clr</td><td>6.643 us</td><td>0.1304 us</td><td>0.1395 us</td><td>6.615 us</td>
-</tr><tr><td>FastReflectionDestructurer</td><td>Clr</td><td>Clr</td><td>11.316 us</td><td>0.2241 us</td><td>0.2201 us</td><td>11.364 us</td>
-</tr><tr><td>OldReflectionDestructurer</td><td>Core</td><td>Core</td><td>807.670 us</td><td>20.8714 us</td><td>51.1980 us</td><td>792.939 us</td>
-</tr><tr><td>CustomDestructurer</td><td>Core</td><td>Core</td><td>785.605 us</td><td>10.6901 us</td><td>9.9995 us</td><td>788.961 us</td>
-</tr><tr><td>FastReflectionDestructurer</td><td>Core</td><td>Core</td><td>883.007 us</td><td>42.0822 us</td><td>124.0805 us</td><td>813.333 us</td>
+</thead><tbody><tr><td>ReflectionDestructurer</td><td>Clr</td><td>Clr</td><td>11.180 us</td><td>0.1812 us</td><td>0.1415 us</td>
+</tr><tr><td>CustomDestructurer</td><td>Clr</td><td>Clr</td><td>6.408 us</td><td>0.0514 us</td><td>0.0456 us</td>
+</tr><tr><td>ReflectionDestructurer</td><td>Core</td><td>Core</td><td>742.092 us</td><td>7.2303 us</td><td>6.0376 us</td>
+</tr><tr><td>CustomDestructurer</td><td>Core</td><td>Core</td><td>750.728 us</td><td>18.9935 us</td><td>34.2492 us</td>
 </tr></tbody></table>
 </body>
 </html>
